Rename "shareable URL" to "public URL"
